docker-swarm/collector/requirements.txt
lxml
mysql-connector-python
requests
